Add SelectRegionFragment, ViewModel, interactor-interactorImpl-repository (TO DO repositryImpl) detekt
