+++ app/src/main/java/ru/practicum/android/diploma/domain/api/FilterInteractor.kt
@@ -3,7 +3,6 @@ package ru.practicum.android.diploma.domain.api
 import kotlinx.coroutines.flow.Flow
 import ru.practicum.android.diploma.domain.models.Country
 import ru.practicum.android.diploma.domain.models.Filter
-import ru.practicum.android.diploma.domain.models.Region
 import ru.practicum.android.diploma.presentation.util.Resource
 
 interface FilterInteractor {
@@ -13,10 +12,4 @@ interface FilterInteractor {
     fun applyCountryFilter(country: Country)
     fun getSelectedCountry(): Country
     fun clearCountryFilter()
-    //
-    fun getRegions(countryId: String): Flow<Resource<List<Region>>>
-    fun applyRegionFilter(region: Region)
-    fun searchRegionByName(regionName: String): Flow<Resource<List<Region>>>
-    fun getSelectedRegion(): Region
-    fun clearRegionFilter()
 }
