+++ app/src/main/java/ru/practicum/android/diploma/domain/api/FilterRepository.kt
@@ -1,17 +1,9 @@
 package ru.practicum.android.diploma.domain.api
 
-import kotlinx.coroutines.flow.Flow
 import ru.practicum.android.diploma.domain.models.Filter
-import ru.practicum.android.diploma.domain.models.Region
-import ru.practicum.android.diploma.presentation.util.Resource
 
 interface FilterRepository {
     fun load(): Filter?
     fun write(filter: Filter)
-    //
-    fun getRegions(countryId: String): Flow<Resource<List<Region>>>
-    fun applyRegionFilter(region: Region)
-    fun searchRegionByName(regionName: String): Flow<Resource<List<Region>>>
-    fun getSelectedRegion(): Region
-    fun clearRegionFilter()
+
 }
